requirements.txt
selenium>=4.1.0
pillow>=9.0.0
webdriver-manager>=3.8.0